ai_scanner/requirements.txt
anthropic==0.42.0
google-genai==0.3.0
playwright==1.49.0
redis[hiredis]==5.2.0
asyncpg==0.30.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
structlog==24.4.0
tenacity==9.0.0
